+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DataManager.extract_solution argument `extract_dvar_names` can also be a `Dict[str, str]` where the value if the solumn column name. 
 - DataManager.extract_solution argument `solution_column_name_post_fix` to allow custom post fix (instead of hard-coded 'Sol')
 - ScenarioRunner logs Excel input file name
+- utils.df_itertuples_with_index_names()
 ### Fixed
 - Small error in ScenarioRunner.insert_outputs_in_db